+++ metrics/__init__.py
@@ -1,3 +1,3 @@
-from ._regression import r2_score
+from ._regression import r2_score, mse
 
-__all__ = ["r2_score"]
+__all__ = ["r2_score", "mse"]